Move league statistics to league_statistics module
