+++ lib/stat_tracker.rb
@@ -3,9 +3,11 @@ require 'pry'
 # require_relative used so program still works when run from spec_harness
 require_relative 'stat_helper'
 require_relative 'team_statistics'
+require_relative 'league_statistics'
 
 class StatTracker < StatHelper
   include TeamStatistics
+  include LeagueStatistics
   attr_reader :games, :teams, :game_teams
   def initialize(games, teams, game_teams)
     @games = games
@@ -97,11 +99,6 @@ class StatTracker < StatHelper
     season_goal_averages
   end
 #-----------------------------------League Statistics-----------------------------------
-  # Original method from iteration 2
-  def count_of_teams
-    @teams.length
-  end
-
   # Helper method used in best_offense, worst_offense - Added as part of I3 refactor
   def avg_goals_per_game
     hash = {}
@@ -120,21 +117,6 @@ class StatTracker < StatHelper
       [team_id, (games_goals_hash[:goals].to_f/games_goals_hash[:games])]
     end
     avg_goals_per_game
-  end
-
-  # Original method from iteration 2
-  def best_offense
-    team_goals_per_game = avg_goals_per_game
-    # Find the team_id and name of the team w/ highest avg goals
-    best_offense_id = team_goals_per_game.max_by {|team_id, avg_goals| avg_goals}[0]
-    @teams.find {|team| team[:team_id] == best_offense_id}[:teamname]
-  end
-
-  # Original method from iteration 2
-  def worst_offense
-    team_goals_per_game = avg_goals_per_game
-    worst_offense_id = team_goals_per_game.min_by {|team_id, avg_goals| avg_goals}[0]
-    @teams.find {|team| team[:team_id] == worst_offense_id}[:teamname]
   end
 
   # Helper method is used in average_scores_for_all_visitors & average_scores_for_all_home_teams
@@ -184,41 +166,6 @@ class StatTracker < StatHelper
     @home_hash
   end
 
-  # Original method from Iteration 2
-  def highest_scoring_visitor
-    average_scores_for_all_visitors
-    highest_scoring_team = @teams.find do |team|
-      team[:team_id].to_i == @visitor_hash.key(@visitor_hash.values.max)
-    end
-    highest_scoring_team[:teamname]
-  end
-
-  # Original method from Iteration 2
-  def highest_scoring_home_team
-    average_scores_for_all_home_teams
-    highest_scoring_team = @teams.find do |team|
-      team[:team_id].to_i == @home_hash.key(@home_hash.values.max)
-    end
-    highest_scoring_team[:teamname]
-  end
-
-  # Original method from Iteration 2
-  def lowest_scoring_visitor
-    average_scores_for_all_visitors
-    lowest_scoring_team = @teams.find do |team|
-      team[:team_id].to_i == @visitor_hash.key(@visitor_hash.values.min)
-    end
-    lowest_scoring_team[:teamname]
-  end
-
-  # Original method from Iteration 2
-  def lowest_scoring_home_team
-    average_scores_for_all_home_teams
-    highest_scoring_team = @teams.find do |team|
-      team[:team_id].to_i == @home_hash.key(@home_hash.values.min)
-    end
-    highest_scoring_team[:teamname]
-  end
 #-----------------------------------Season Statistics-----------------------------------
   # Helper method is used in most_accurate_team & least_accurate_team
   # Returns an array of all @game_teams rows from a given season
