+++ spec/spec_helper.rb
@@ -4,6 +4,7 @@ SimpleCov.start
 require_relative 'stat_helper_spec'
 require_relative 'stat_tracker_spec'
 require_relative 'team_statistics_spec'
+require_relative 'league_statistics_spec'
